remove monitor.py
