+++ field_friend/interface/components/__init__.py
@@ -1,8 +1,8 @@
 from .camera_card import CameraCard
 from .header import create_header
+from .key_controls import KeyControls
 from .leaflet_map import LeafletMap
 from .log_monitor import LogMonitor
-from .monitoring import Monitoring
 from .operation import Operation
 from .path_object import PathObject as path_object
 from .robot_scene import RobotScene
@@ -10,9 +10,9 @@ from .status_drawer import create_status_drawer
 
 __all__ = [
     'CameraCard',
+    'KeyControls',
     'LeafletMap',
     'LogMonitor',
-    'Monitoring',
     'Operation',
     'RobotScene',
     'create_header',
